os: xtrans: silence printf format warning on prmsg() call
size_t is unsigned, but might have different sizes depending on CPU arch,
so casting it to unsigned long, so we can safely use %ld.

+++ os/Xtranssock.c
@@ -1340,7 +1340,7 @@ static int _XSERVTransSocketRead (
 static ssize_t _XSERVTransSocketWritev (
     XtransConnInfo ciptr, struct iovec *iov, size_t iovcnt)
 {
-    prmsg (2,"SocketWritev(%d,%p,%d)\n", ciptr->fd, (void *) iov, iovcnt);
+    prmsg (2,"SocketWritev(%d,%p,%ld)\n", ciptr->fd, (void *) iov, (unsigned long)iovcnt);
 
 #if XTRANS_SEND_FDS
     if (ciptr->send_fds)
